🧪 Move utility file to fix build error and pass tests
Moved `svg-to-png-utils.js` from `src/pages/image/utils/` to `src/utils/image/`
to fix Astro build treating it as a page and throwing `No API Route handler
exists for the method "GET"` error. Updated tests and imports accordingly.
