+++ src/pages/image/svg-to-png.astro
@@ -129,7 +129,7 @@ const faqItems = [
   </style>
 
   <script>
-    import { formatSize, handleFile as processFile, convertImage as doConvertImage } from './utils/svg-to-png-utils.js';
+    import { formatSize, handleFile as processFile, convertImage as doConvertImage } from '../../utils/image/svg-to-png-utils.js';
 
     const dropZone = document.getElementById('dropZone');
     const fileInput = document.getElementById('fileInput');
