+++ src/utils/image/svg-to-png-utils.js
